Split Sun UV data config
