+++ scripts/quality-baseline.json
@@ -2,5 +2,5 @@
   "inlineEventAttributes": 0,
   "windowReferences": 202,
   "largeJsFilesOver800Lines": 24,
-  "maxJsFileLines": 1178
+  "maxJsFileLines": 1174
 }
