+++ tsconfig.json
@@ -245,6 +245,7 @@
     "js/sun-session-ui.js",
     "js/sun-sessions-store.js",
     "js/sun-spectrum.js",
+    "js/sun-uvdata-config.js",
     "js/sun-uvdata.js",
     "js/sun.js",
     "js/sync-actions.js",
